refactor: 컴포넌트 파일 정리
