+++ styles/globals.css
@@ -14,7 +14,7 @@ a {
 }
 
 * {
-  max-width: 1600px;
+  max-width: 1200px;
   box-sizing: border-box;
 }
